+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Exception `InvalidQueryIdException`
 - Exception `ChatDescriptionIsNotModifiedException`
 - Exception `ChatNotModifiedException`
+- Exception `SendMessageToBotException`
 
 ### Changed